All Order policies added

+++ app/Http/Controllers/OrderController.php
@@ -7,6 +7,7 @@ use App\Http\Resources\OrderResource;
 use App\Models\Order;
 use App\Models\Product;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Gate;
 
 class OrderController extends Controller
 {
@@ -16,8 +17,15 @@ class OrderController extends Controller
      */
     public function index(): JsonResponse
     {
-        $orderList = OrderResource::collection(Order::with(['user', 'products'])->get());
-        return response()->json($orderList, 201);
+        $response = Gate::inspect('viewAny', Order::class);
+
+        if ($response->allowed()) {
+            $orderList = OrderResource::collection(Order::with(['user', 'products'])->get());
+            return response()->json($orderList, 201);
+        }
+        else {
+            return response()->json(['message' => $response->message()], 403);
+        }
     }
 
     /**
@@ -28,8 +36,16 @@ class OrderController extends Controller
     public function getByUserId($id): JsonResponse
     {
         $order = Order::where('user_id', $id)->with(['user', 'products'])->get();
-        $order = OrderResource::collection($order);
-        return response()->json($order, 201);
+
+        $response = Gate::inspect('view', $order->first());
+
+        if ($response->allowed()) {
+            $order = OrderResource::collection($order);
+            return response()->json($order, 201);
+        }
+        else {
+            return response()->json(['message' => $response->message()], 403);
+        }
     }
 
     /**
@@ -40,8 +56,16 @@ class OrderController extends Controller
     public function getById($id): JsonResponse
     {
         $order = Order::where('id', $id)->with(['user', 'products'])->get();
-        $order = OrderResource::collection($order);
-        return response()->json($order, 201);
+
+        $response = Gate::inspect('view', $order->first());
+
+        if ($response->allowed()) {
+            $order = OrderResource::collection($order);
+            return response()->json($order, 201);
+        }
+        else {
+            return response()->json(['message' => $response->message()], 403);
+        }
     }
 
     /**
@@ -51,36 +75,42 @@ class OrderController extends Controller
      */
     public function store(StoreOrderRequest $request): JsonResponse
     {
-        $highestOrderNumber = Order::max('order_number');
-        $newOrderNumber = $highestOrderNumber ? $highestOrderNumber + 1 : 1;
+        $response = Gate::inspect('create', Order::class);
 
-        $order = Order::create([
-            'user_id' => $request->user()->id,
-            'order_number' => $newOrderNumber,
-            'price' => $request->input('price'),
-            'validatedStatus' => $request->input('validatedStatus'),
-            'shippingCountry' => $request->input('shippingCountry'),
-            'shippingMode' => $request->input('shippingMode'),
-            'shippingPrice' => $request->input('shippingPrice'),
-            'creatorCode' => $request->input('creatorCode'),
-            'promoCode' => $request->input('promoCode'),
-        ]);
+        if($response->allowed()){
+            $highestOrderNumber = Order::max('order_number');
+            $newOrderNumber = $highestOrderNumber ? $highestOrderNumber + 1 : 1;
 
-        foreach ($request->input('products') as $product) {
-            $order->products()->attach(
-                $product['id'],
-                ['quantity' => $product['quantity'],
-                    'price' => $product['price'],
-                    'size' => array_key_exists('size', $product) ? $product['size'] : null,
-                    'color' => array_key_exists('color', $product) ? $product['color'] : null,
-                    'material' => array_key_exists('material', $product) ? $product['material'] : null,
-                ]
-            );
-            Product::find($product['id'])->decrement('stock', $product['quantity']);
+            $order = Order::create([
+                'user_id' => $request->user()->id,
+                'order_number' => $newOrderNumber,
+                'price' => $request->input('price'),
+                'validatedStatus' => $request->input('validatedStatus'),
+                'shippingCountry' => $request->input('shippingCountry'),
+                'shippingMode' => $request->input('shippingMode'),
+                'shippingPrice' => $request->input('shippingPrice'),
+                'creatorCode' => $request->input('creatorCode'),
+                'promoCode' => $request->input('promoCode'),
+            ]);
+
+            foreach ($request->input('products') as $product) {
+                $order->products()->attach(
+                    $product['id'],
+                    ['quantity' => $product['quantity'],
+                        'price' => $product['price'],
+                        'size' => array_key_exists('size', $product) ? $product['size'] : null,
+                        'color' => array_key_exists('color', $product) ? $product['color'] : null,
+                        'material' => array_key_exists('material', $product) ? $product['material'] : null,
+                    ]
+                );
+                Product::find($product['id'])->decrement('stock', $product['quantity']);
+            }
+            $order->save();
+            return response()->json($order, 201);
         }
-
-        $order->save();
-        return response()->json($order, 201);
+       else {
+            return response()->json(['message' => $response->message()], 403);
+       }
     }
 
     /**
@@ -91,8 +121,16 @@ class OrderController extends Controller
     public function destroy($id): JsonResponse
     {
         $order = Order::find($id);
-        $order->orderProducts()->delete();
-        $order->delete();
-        return response()->json(['message' => 'Order deleted'], 201);
+
+        $response = Gate::inspect('delete', $order);
+
+        if ($response->allowed()) {
+            $order->orderProducts()->delete();
+            $order->delete();
+            return response()->json(['message' => 'Order deleted'], 201);
+        }
+        else {
+            return response()->json(['message' => $response->message()], 403);
+        }
     }
 }
